# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProviEmplea API",
        "description": "API REST para la plataforma ProviEmplea.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
        "/admin/contactos": {
            "get": {
                "tags": [
                    "Administración"
                ],
                "summary": "Listar contactos solicitados",
                "description": "Retorna las solicitudes de contacto con empresa y persona asociadas.",
                "operationId": "a5f060bc7b6fa5377c1497af4ddcd2d7",
                "responses": {
                    "200": {
                        "description": "Listado de contactos obtenido correctamente"
                    }
                }
            },
            "post": {
                "tags": [
                    "Administración"
                ],
                "summary": "Crear contacto solicitado",
                "description": "Registra una solicitud de contacto entre una empresa y una persona.",
                "operationId": "5dbc342d1c62d99c382fd63d39ff2d4f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "empresa_id": 1,
                                "persona_id": 1,
                                "estado": "pendiente",
                                "notas_admin": "Empresa interesada en contactar al talento"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contacto creado correctamente"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                }
            }
        },
        "/admin/contactos/{id}": {
            "patch": {
                "tags": [
                    "Administración"
                ],
                "summary": "Actualizar contacto solicitado",
                "description": "Actualiza el estado o notas de una solicitud de contacto.",
                "operationId": "5783d5aacdc283ea7d5688d1124aa8d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "estado": "contactado",
                                "notas_admin": "Se realizó el primer contacto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contacto actualizado correctamente"
                    },
                    "404": {
                        "description": "Contacto no encontrado"
                    }
                }
            }
        },
...
        "/empresas": {
            "get": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Listar empresas",
                "description": "Retorna el listado de empresas registradas.",
                "operationId": "22631014ece23858d19a7ef1c89f0e78",
                "responses": {
                    "200": {
                        "description": "Listado de empresas obtenido correctamente"
                    }
                }
            },
            "post": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Crear empresa",
                "description": "Registra una nueva empresa.",
                "operationId": "edd395f6abdaa55cb0284b3b198d7ad6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "nombre_empresa": "TechCorp SpA",
                                "rut_empresa": "76123456-7",
                                "email": "rrhh@techcorp.cl",
                                "tipo_empresa": "contratacion-directa",
                                "rubro": "Tecnología",
                                "beneficios": [
                                    "Seguro complementario",
                                    "Trabajo remoto"
                                ],
                                "contacto_nombre": "Ana López",
                                "contacto_email": "ana@techcorp.cl",
                                "validado": false,
                                "activo": true
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empresa creada correctamente"
                    }
                }
            }
        },
        "/empresas/{id}": {
            "get": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Obtener empresa por ID",
                "operationId": "1ab6ccd465b58367f4064aa417a103be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa encontrada"
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Actualizar empresa",
                "operationId": "b61661cfb086a49afd6ebebf15ed6011",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa actualizada correctamente"
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Eliminar empresa",
                "operationId": "56b2aac859ceb472597b45436c7f5f12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa eliminada correctamente"
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    }
                }
            }
        },
        "/empresas/{id}/validar": {
            "patch": {
                "tags": [
                    "Empresas"
                ],
                "summary": "Validar empresa",
                "operationId": "8e0e700c208ef6231bdf7ccd0bf869d5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empresa validada correctamente"
                    },
                    "404": {
                        "description": "Empresa no encontrada"
                    }
                }
            }
        },
...
    },
...
}